pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-jarvis1net-py"
version = "2.0.1"
description = "MCP server: filesystem, shell diagnostics, Microsoft Graph (Python, stdio)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.0.0",
  "llama-index-core>=0.11.0",
  "llama-index-embeddings-openai>=0.2.0",
  "llama-index-vector-stores-qdrant>=0.3.0",
  "qdrant-client>=1.11.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.2",
  "requests>=2.31.0",
]

[project.scripts]
mcp-jarvis1net-py = "server:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["server"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"rag" = ["rag_config.json", "sources/*.yaml", "tests/*.json"]

